use Float64Array

+++ src/lib/synthesis/brownNoise.ts
@@ -5,7 +5,7 @@ export const brownNoise = (
   whiteNoiseGenerator = whiteNoise,
   state = 0
 ) => {
-  const buffer = new Float32Array(lengthInSamples)
+  const buffer = new Float64Array(lengthInSamples)
 
   for (let i = 0; i < lengthInSamples; i++) {
     const white = whiteNoiseGenerator()
